Issue history UI enhancements
