+++ epictrack-web/src/components/icons/index.tsx
@@ -415,6 +415,25 @@ const LockOpenIcon = (props: IconProps) => {
   );
 };
 
+const IssueHistoryIcon = (props: IconProps) => {
+  return (
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      width="50"
+      height="45"
+      viewBox="0 0 50 45"
+      fill="none"
+    >
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M4.08029 6.09127C5.48459 5.8063 6.54157 4.56478 6.54157 3.07639C6.54157 1.37735 5.16423 0 3.46518 0C1.76614 0 0.388794 1.37735 0.388794 3.07639C0.388794 4.56466 1.44561 5.8061 2.84973 6.0912V17.2891C1.44561 17.5742 0.388794 18.8157 0.388794 20.3039C0.388794 21.7922 1.44561 23.0336 2.84973 23.3187V34.5176C1.44561 34.8027 0.388794 36.0442 0.388794 37.5324C0.388794 39.2315 1.76614 40.6088 3.46518 40.6088C5.16423 40.6088 6.54157 39.2315 6.54157 37.5324C6.54157 36.0441 5.48459 34.8025 4.08029 34.5176V23.3188C5.48459 23.0338 6.54157 21.7923 6.54157 20.3039C6.54157 18.8155 5.48459 17.574 4.08029 17.289V6.09127ZM11.4639 1.23056C11.4639 0.550938 12.0148 0 12.6944 0H48.3805C49.0602 0 49.6111 0.550938 49.6111 1.23056C49.6111 1.91017 49.0602 2.46111 48.3806 2.46111H12.6944C12.0148 2.46111 11.4639 1.91017 11.4639 1.23056ZM12.6945 17.2275C12.0149 17.2275 11.464 17.7785 11.464 18.4581C11.464 19.1377 12.0149 19.6887 12.6945 19.6887H48.3807C49.0603 19.6887 49.6112 19.1377 49.6112 18.4581C49.6112 17.7785 49.0603 17.2275 48.3807 17.2275H12.6945ZM11.464 35.6866C11.464 35.007 12.0149 34.4561 12.6945 34.4561H48.3807C49.0603 34.4561 49.6112 35.007 49.6112 35.6866C49.6112 36.3662 49.0603 36.9172 48.3807 36.9172H12.6945C12.0149 36.9172 11.464 36.3662 11.464 35.6866ZM12.6945 3.69141C12.0149 3.69141 11.464 4.24234 11.464 4.92196C11.464 5.60158 12.0149 6.15252 12.6945 6.15252H48.3807C49.0603 6.15252 49.6112 5.60158 49.6112 4.92196C49.6112 4.24234 49.0603 3.69141 48.3807 3.69141H12.6945ZM11.464 22.1495C11.464 21.4699 12.0149 20.9189 12.6945 20.9189H48.3807C49.0603 20.9189 49.6112 21.4699 49.6112 22.1495C49.6112 22.8291 49.0603 23.3801 48.3807 23.3801H12.6945C12.0149 23.3801 11.464 22.8291 11.464 22.1495ZM12.6945 38.1475C12.0149 38.1475 11.464 38.6984 11.464 39.378C11.464 40.0576 12.0149 40.6086 12.6945 40.6086H48.3807C49.0603 40.6086 49.6112 40.0576 49.6112 39.378C49.6112 38.6984 49.0603 38.1475 48.3807 38.1475H12.6945ZM11.464 8.61337C11.464 7.93375 12.0149 7.38281 12.6945 7.38281H48.3807C49.0603 7.38281 49.6112 7.93375 49.6112 8.61337C49.6112 9.29298 49.0603 9.84392 48.3807 9.84392H12.6945C12.0149 9.84392 11.464 9.29298 11.464 8.61337ZM12.6944 24.6111C12.0148 24.6111 11.4639 25.162 11.4639 25.8417C11.4639 26.5213 12.0148 27.0722 12.6944 27.0722H48.3806C49.0602 27.0722 49.6111 26.5213 49.6111 25.8417C49.6111 25.162 49.0602 24.6111 48.3805 24.6111H12.6944ZM11.4639 43.0695C11.4639 42.3899 12.0148 41.839 12.6944 41.839H48.3805C49.0602 41.839 49.6111 42.3899 49.6111 43.0695C49.6111 43.7491 49.0602 44.3001 48.3806 44.3001H12.6944C12.0148 44.3001 11.4639 43.7491 11.4639 43.0695Z"
+        fill="#F2F2F2"
+      />
+    </svg>
+  );
+};
+
 const icons: { [x: string]: React.FC<IconProps> } = {
   AllIcon,
   DashboardIcon,
@@ -458,6 +477,7 @@ const icons: { [x: string]: React.FC<IconProps> } = {
   CloseXIcon,
   LockClosedIcon,
   LockOpenIcon,
+  IssueHistoryIcon,
 };
 
 export default icons;
