+++ epictrack-web/src/components/workPlan/issues/ReportsPreview/ReferralSchedule.tsx
@@ -29,7 +29,15 @@ export const ReferralSchedule = () => {
   );
 
   const issueUpdates = activeApprovedIssues
-    .map((issue) => issue.updates.find((update) => update.is_approved))
+    .map((issue) => {
+      const update = issue.updates.find((update) => update.is_approved);
+      if (!update) return null;
+
+      return {
+        issueTitle: issue.title,
+        ...update,
+      };
+    })
     .filter((update) => Boolean(update));
 
   const latestIssue = activeApprovedIssues?.[0];
@@ -85,9 +93,20 @@ export const ReferralSchedule = () => {
               <Then>
                 <Stack spacing={2} direction="column">
                   {issueUpdates.map((issueUpdate) => (
-                    <ETPreviewText color={Palette.neutral.dark}>
-                      {issueUpdate?.description}
-                    </ETPreviewText>
+                    <>
+                      <ETPreviewText
+                        key={`title-${issueUpdate?.id}`}
+                        color={Palette.neutral.dark}
+                      >
+                        {issueUpdate?.issueTitle}
+                      </ETPreviewText>
+                      <ETPreviewText
+                        key={`description-${issueUpdate?.id}`}
+                        color={Palette.neutral.dark}
+                      >
+                        {issueUpdate?.description}
+                      </ETPreviewText>
+                    </>
                   ))}
                 </Stack>
               </Then>
